Refactor: Move Python migration tests to Wpf3Tests
