+++ test/DynamoCoreWpf3Tests/PythonNodeCustomizationTests.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Controls.Primitives;
@@ -10,6 +12,7 @@ using Autodesk.DesignScript.Runtime;
 using CoreNodeModels;
 using Dynamo.Controls;
 using Dynamo.DocumentationBrowser;
+using Dynamo.Wpf.Extensions;
 using Dynamo.Graph.Workspaces;
 using Dynamo.PythonServices;
 using Dynamo.PythonServices.EventHandlers;
@@ -78,6 +81,72 @@ namespace DynamoCoreWpfTests
             libraries.Add("VMDataBridge.dll");
             libraries.Add("DSCoreNodes.dll");
             base.GetLibrariesToPreload(libraries);
+        }
+
+        [Test]
+        public void NestedCPythonCustomNodes_AreAutoMigratedToPythonNet3_AndEvaluateTo20()
+        {
+            // The auto-migration behavior lives in PythonMigrationViewExtension and is driven by workspace events.
+            // Ensure the DynamoView is marked as loaded and the extension is present.
+            RaiseLoadedEvent(View);
+            DispatcherUtil.DoEvents();
+
+            EnsurePythonMigrationViewExtensionLoaded();
+
+            // Load custom node definitions into the manager first.
+            var pythonDir = Path.Combine(TestDirectory, "core", "python");
+            var childPath = Path.Combine(pythonDir, "CNWithCPython_Child.dyf");
+            var parentPath = Path.Combine(pythonDir, "CNWithCPython_Parent.dyf");
+
+            Assert.IsTrue(File.Exists(childPath), "Missing test file: " + childPath);
+            Assert.IsTrue(File.Exists(parentPath), "Missing test file: " + parentPath);
+
+            Assert.IsTrue(ViewModel.Model.CustomNodeManager.AddUninitializedCustomNode(childPath, true, out _));
+            Assert.IsTrue(ViewModel.Model.CustomNodeManager.AddUninitializedCustomNode(parentPath, true, out _));
+
+            // Open graph and run.
+            Open(@"core\python\WithNestedCPythonCustomNodes.dyn");
+            DispatcherUtil.DoEvents();
+
+            Run();
+            DispatcherUtil.DoEvents();
+
+            // Assert watch value is "20".
+            var watch = Model.CurrentWorkspace.NodeFromWorkspace<Watch>("3b2be8477f5a4ec5a6dc23d9f88a7b7e");
+            Assert.IsNotNull(watch);
+            Assert.AreEqual("20", watch.CachedValue?.ToString());
+        }
+
+        private void EnsurePythonMigrationViewExtensionLoaded()
+        {
+            // If already loaded by the view extension loader, do nothing.
+            if (View.viewExtensionManager.ViewExtensions.Any(e => e != null && e.Name == "Python Migration"))
+            {
+                return;
+            }
+
+            // Load the extension assembly (it is referenced by the test project and should be copied to output).
+            var binDir = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location) ?? string.Empty;
+            var extDllPath = Path.Combine(binDir, "PythonMigrationViewExtension.dll");
+            Assert.IsTrue(File.Exists(extDllPath), "PythonMigrationViewExtension.dll not found next to test binaries: " + extDllPath);
+
+            var asm = Assembly.LoadFrom(extDllPath);
+            var extType = asm.GetType("Dynamo.PythonMigration.PythonMigrationViewExtension", throwOnError: true);
+
+            // The view extension class is internal; instantiate via reflection.
+            var ext = Activator.CreateInstance(extType, nonPublic: true) as IViewExtension;
+            Assert.IsNotNull(ext, "Failed to create PythonMigrationViewExtension via reflection.");
+
+            // Startup + Loaded need internal constructors, but DynamoCoreWpf exposes InternalsVisibleTo to this test assembly.
+            var startupParams = new ViewStartupParams(ViewModel);
+            ext.Startup(startupParams);
+
+            var loadedParams = new ViewLoadedParams(View, ViewModel);
+            ext.Loaded(loadedParams);
+
+            View.viewExtensionManager.Add(ext);
+
+            Assert.IsTrue(View.viewExtensionManager.ViewExtensions.Any(e => e != null && e.Name == "Python Migration"));
         }
 
         /// <summary>
